# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CASCADE API",
    "description": "Synthetic disruption-recovery demonstration API.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/runs": {
      "post": {
        "tags": [
          "workflow"
        ],
        "summary": "Create Run",
        "operationId": "create_run_api_runs_post",
        "parameters": [
          {
            "name": "mode",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/RunMode"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Overrides the body mode field.",
              "title": "Mode"
            },
            "description": "Overrides the body mode field."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRunRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunCreated"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/runs/{run_id}/events": {
      "get": {
        "tags": [
          "workflow"
        ],
        "summary": "Stream Run",
        "operationId": "stream_run_api_runs__run_id__events_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}": {
      "get": {
        "tags": [
          "workflow"
        ],
        "summary": "Get Run",
        "operationId": "get_run_api_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowState"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/dispute-resolution": {
      "post": {
        "tags": [
          "workflow"
        ],
        "summary": "Resolve Dispute",
        "operationId": "resolve_dispute_api_runs__run_id__dispute_resolution_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DisputeResolutionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowState"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/approval": {
      "post": {
        "tags": [
          "workflow"
        ],
        "summary": "Decide Approval",
        "operationId": "decide_approval_api_runs__run_id__approval_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowState"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/benchmarks": {
      "post": {
        "tags": [
          "benchmark"
        ],
        "summary": "Create Benchmark",
        "operationId": "create_benchmark_api_benchmarks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBenchmarkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BenchmarkCreated"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/benchmarks/{benchmark_id}/events": {
      "get": {
        "tags": [
          "benchmark"
        ],
        "summary": "Stream Benchmark",
        "operationId": "stream_benchmark_api_benchmarks__benchmark_id__events_get",
        "parameters": [
          {
            "name": "benchmark_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Benchmark Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/benchmarks/{benchmark_id}": {
      "get": {
        "tags": [
          "benchmark"
        ],
        "summary": "Get Benchmark",
        "operationId": "get_benchmark_api_benchmarks__benchmark_id__get",
        "parameters": [
          {
            "name": "benchmark_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Benchmark Id"
            }
          },
          {
            "name": "since",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Skip events already received over SSE.",
              "default": 0,
              "title": "Since"
            },
            "description": "Skip events already received over SSE."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BenchmarkState"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AgentActivity": {
        "properties": {
          "agent": {
            "$ref": "#/components/schemas/AgentName"
          },
          "objective": {
            "type": "string",
            "title": "Objective"
          },
          "status": {
            "$ref": "#/components/schemas/AgentStatus",
            "default": "WAITING"
          },
          "confidence": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Confidence"
              },
              {
                "type": "null"
              }
            ]
          },
          "last_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Summary"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "agent",
          "objective"
        ],
        "title": "AgentActivity"
      },
...
      "ApprovalDecision": {
        "type": "string",
        "enum": [
          "APPROVED",
          "REJECTED"
        ],
        "title": "ApprovalDecision"
      },
      "ApprovalRequest": {
        "properties": {
          "plan_archetype": {
            "$ref": "#/components/schemas/PlanArchetype"
          },
          "decision": {
            "$ref": "#/components/schemas/ApprovalDecision"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "plan_archetype",
          "decision"
        ],
        "title": "ApprovalRequest"
      },
...
      "BenchmarkCreated": {
        "properties": {
          "benchmark_id": {
            "type": "string",
            "title": "Benchmark Id"
          },
          "stage": {
            "$ref": "#/components/schemas/BenchmarkStage"
          },
          "events_url": {
            "type": "string",
            "title": "Events Url"
          },
          "playback_notice": {
            "type": "string",
            "title": "Playback Notice"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "benchmark_id",
          "stage",
          "events_url",
          "playback_notice"
        ],
        "title": "BenchmarkCreated"
      },
...
      "BenchmarkState": {
        "properties": {
          "benchmark_id": {
            "type": "string",
            "title": "Benchmark Id"
          },
          "stage": {
            "$ref": "#/components/schemas/BenchmarkStage"
          },
          "config": {
            "$ref": "#/components/schemas/BenchmarkConfig"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/BenchmarkEvent"
            },
            "type": "array",
            "title": "Events"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BenchmarkResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "playback_notice": {
            "type": "string",
            "title": "Playback Notice"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "benchmark_id",
          "stage",
          "config",
          "playback_notice"
        ],
        "title": "BenchmarkState"
      },
...
      "CreateBenchmarkRequest": {
        "properties": {
          "seed": {
            "type": "integer",
            "title": "Seed",
            "default": 42
          },
          "arms": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FleetArm"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Arms"
          },
          "brain": {
            "$ref": "#/components/schemas/FleetBrainMode",
            "default": "SCRIPTED"
          },
          "playback_speed": {
            "type": "number",
            "maximum": 100.0,
            "exclusiveMinimum": 0.0,
            "title": "Playback Speed",
            "default": 1.0
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "CreateBenchmarkRequest"
      },
      "CreateRunRequest": {
        "properties": {
          "delay_hours": {
            "type": "integer",
            "maximum": 24.0,
            "minimum": 6.0,
            "title": "Delay Hours",
            "default": 18
          },
          "priority_emphasis": {
            "$ref": "#/components/schemas/PriorityEmphasis",
            "default": "BALANCED"
          },
          "alternative_sailing_failure": {
            "type": "boolean",
            "title": "Alternative Sailing Failure",
            "default": true
          },
          "mode": {
            "$ref": "#/components/schemas/RunMode",
            "default": "LIVE_STUB"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "CreateRunRequest",
        "description": "Scenario controls plus explicit run-mode selection."
      },
...
      "DisputeResolutionRequest": {
        "properties": {
          "dispute_id": {
            "type": "string",
            "title": "Dispute Id"
          },
          "confirmed_constraint": {
            "type": "string",
            "title": "Confirmed Constraint"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "dispute_id",
          "confirmed_constraint"
        ],
        "title": "DisputeResolutionRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PlanArchetype": {
        "type": "string",
        "enum": [
          "AGGRESSIVE_RUSH",
          "STANDARD_REBOOK",
          "OPTIMIZED_HYBRID"
        ],
        "title": "PlanArchetype"
      },
...
      "PriorityEmphasis": {
        "type": "string",
        "enum": [
          "CARGO_PROTECTION",
          "CONGESTION_REDUCTION",
          "BALANCED"
        ],
        "title": "PriorityEmphasis"
      },
...
      "RunCreated": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "mode": {
            "$ref": "#/components/schemas/RunMode"
          },
          "stage": {
            "$ref": "#/components/schemas/WorkflowStage"
          },
          "events_url": {
            "type": "string",
            "title": "Events Url"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run_id",
          "mode",
          "stage",
          "events_url"
        ],
        "title": "RunCreated"
      },
      "RunMode": {
        "type": "string",
        "enum": [
          "LIVE_STUB",
          "LIVE_GEMINI",
          "LIVE_CLAUDE",
          "DEMO_REPLAY"
        ],
        "title": "RunMode"
      },
...
      "ScenarioState": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "alert": {
            "$ref": "#/components/schemas/DisruptionAlert"
          },
          "objective": {
            "type": "string",
            "title": "Objective"
          },
          "planning_horizon_hours": {
            "type": "integer",
            "title": "Planning Horizon Hours"
          },
          "controls": {
            "$ref": "#/components/schemas/ScenarioControls"
          },
          "synthetic_notice": {
            "type": "string",
            "title": "Synthetic Notice"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "description",
          "alert",
          "objective",
          "planning_horizon_hours",
          "controls",
          "synthetic_notice"
        ],
        "title": "ScenarioState"
      },
...
      "TraceEvent": {
        "properties": {
          "event_id": {
            "type": "string",
            "title": "Event Id"
          },
          "sequence": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Sequence"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "kind": {
            "$ref": "#/components/schemas/EventKind"
          },
          "stage": {
            "$ref": "#/components/schemas/WorkflowStage"
          },
          "agent": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AgentName"
              },
              {
                "type": "null"
              }
            ]
          },
          "objective": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Objective"
          },
          "input_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Summary"
          },
          "tool": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tool"
          },
          "decision_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision Summary"
          },
          "confidence": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Confidence"
              },
              {
                "type": "null"
              }
            ]
          },
          "assumptions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Assumptions"
          },
          "result": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "elapsed_ms": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Elapsed Ms"
          },
          "next_handoff": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AgentName"
              },
              {
                "type": "null"
              }
            ]
          },
          "parallel_group": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parallel Group"
          },
          "model_exchanges": {
            "items": {
              "$ref": "#/components/schemas/ModelExchange"
            },
            "type": "array",
            "title": "Model Exchanges"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "event_id",
          "sequence",
          "timestamp",
          "kind",
          "stage"
        ],
        "title": "TraceEvent"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WorkflowStage": {
        "type": "string",
        "enum": [
          "READY",
          "ASSESSING",
          "DISPUTE",
          "PLANNING",
          "AWAITING_APPROVAL",
          "EXECUTING",
          "COMPLETE",
          "FAILED"
        ],
        "title": "WorkflowStage"
      },
      "WorkflowState": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "mode": {
            "$ref": "#/components/schemas/RunMode"
          },
          "stage": {
            "$ref": "#/components/schemas/WorkflowStage"
          },
          "scenario": {
            "$ref": "#/components/schemas/ScenarioState"
          },
          "activities": {
            "items": {
              "$ref": "#/components/schemas/AgentActivity"
            },
            "type": "array",
            "title": "Activities"
          },
          "trace": {
            "items": {
              "$ref": "#/components/schemas/TraceEvent"
            },
            "type": "array",
            "title": "Trace"
          },
          "active_dispute": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Dispute"
              },
              {
                "type": "null"
              }
            ]
          },
          "results": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunResults"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run_id",
          "mode",
          "stage",
          "scenario",
          "activities"
        ],
        "title": "WorkflowState"
      },
...
    }
  }
}